refactor getting PDF text
